requirements.txt
# LegalEagle - Python Dependencies
# Latest versions compatible with langchain 1.0+

# ========== Core Web Framework ==========
Flask>=3.0.0
flask-cors>=4.0.0
Werkzeug>=3.0.0

# ========== Environment & Configuration ==========
python-dotenv>=1.0.0

# ========== Authentication & Security ==========
PyJWT>=2.8.0
bcrypt>=4.1.0

# ========== Database ==========
SQLAlchemy>=2.0.0

# ========== LangChain Core & AI (Latest Versions) ==========
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0

# ========== Vector Database ==========
chromadb>=0.5.0

# ========== OpenAI ==========
openai>=1.50.0
tiktoken>=0.8.0

# ========== Optional LLM Providers ==========
langchain-google-genai>=2.0.0
langchain-anthropic>=0.1.0
langchain-groq>=0.1.0

# ========== Document Loaders ==========
pymupdf>=1.23.0
pypdf>=4.0.0
docx2txt>=0.8
python-docx>=1.0.0
unstructured>=0.12.0

# ========== Embeddings (HuggingFace) ==========
sentence-transformers>=2.2.0

# ========== Additional Utilities ==========
requests>=2.31.0
